requirements.txt
langchain_core==1.0.7
langchain_ollama==1.0.0
langchain_openai==1.0.3
langgraph==1.0.3
numpy==2.3.5
psycopg2_binary==2.9.11
pydantic==2.12.4
qdrant_client==1.16.0
rapidfuzz==3.14.3
sentence_transformers==5.1.2
